Reimplemented program using RageAgainstThePixel's OpenAI wrapper

+++ Assets/_Game/Scripts/Gameplay/Open AI Integration/Vector Search/IVectorObject.cs
@@ -6,6 +6,6 @@ namespace ChatGPT_Detective
 {
     public interface IVectorObject
     {
-        public float[] GetVector();
+        public double[] GetVector();
     }
 }
